Refactor: [T-28] import문 수정 및 폴더명 변경
https://www.notion.so/T-28-Guide-page-734de37c67b14787aa56efe4ef7ea077?pvs=4
