+++ src/components/DescriptionWithCode/style.scss
@@ -7,5 +7,6 @@
   .text {
     padding-bottom: 18px;
     font-size: 2rem;
+    white-space: pre-wrap;
   }
 }
